backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
easyocr==1.7.1
Pillow==10.2.0
google-generativeai==0.4.0
python-dotenv==1.0.1
aiofiles==23.2.1
bcrypt==4.1.2
